Add a flag to optionally use the ghc-paths library (off by default).
Adds a new flag, "use-ghc-paths", which adds the `ghc-paths` library as a
dependency of ghc-paths and uses it to locate the GHC libdir.

This functionality used to be present, but was removed a few years ago:
b1af2bcf91163269983de0914f0dd54912c2db0b
Since then the code has been refactored so that the change is smaller.

The flag is off by default, since it makes the binaries not prefix-independent.
More specifically, it embeds in them a full, absolute path to the GHC libdir.