Fix suprising behavior of 'withReset' and friends
Before this commit, 'withReset' would behave very surprisingly in
combination with 'simulate' or 'sample':

    >>> rst0 = fromList [True, True, False, False, True, True]
    >>> rst1 = unsafeFromHighPolarity @System rst0
    >>> sig = withReset rst1 (register 'a' (pure 'b'))
    >>> sampleN @System 6 sig
    "aabbbb"

Even though rst1 is asserted on the 5th and 6th clock cycle, we never
see the reset value, a, after the initial two. But even the first two
as are suspicious. We'd expect to see three as (one powerup, two
reset) instead! It seems that 'sampleN' is generating a default reset
value and using that one. Indeed, when we inspect the type of sig we
see it is as we never applied the reset at all:

    >>> :t sig
    (HiddenReset dom, ..) => Signal dom Char

To figure out what's going on, let's inspect the type signature of
'withReset':

    withReset
      :: KnownDomain dom
      => Reset dom
      -> (HiddenReset dom => r)
      -> r

It turns out that GHC is perfectly happy to /only/ resolve r when
given the following as its second argument:

    reg :: (HiddenReset dom, ..) => Signal dom Char
    reg = register 'a' (pure 'b')

i.e., withReset would conceptually resolve to something like:

    withReset*
      :: forall dom r
       . KnownDomain dom
      => Reset dom
      -> (HiddenReset dom  => (HiddenReset dom0 => Signal dom0 Char))
      -> (HiddenReset dom0 => Signal dom0 Char)

thus completely discarding its given reset, and leaving it up to
sample to actually generate the reset. The reason GHC is free to do
this, is that nothing in r ties dom to the "dom inside r" simply
because it's not mentioned in the type signature.

This commit adds a few variants of 'withReset' that explicitly take
an argument, thus "tying" the `dom` in that argument to the result of
the function given to it.

'withReset0' behaves exactly like we expect it would:

    >>> rst0 = fromList [True, True, False, False, True, True]
    >>> rst1 = unsafeFromHighPolarity rst0
    >>> sig = withReset0 rst1 (register 'a' (pure 'b'))
    >>> sampleN @System 6 sig
    "aaabaa"