Add type classes to reliably use hidden constructs
This is a follow-up of PR #655. That PR identified unexpected behavior
when using 'withReset' in combination with 'sample'. It originally
proposed a solution with a type class "HasDomain" that solved the
unexpected behavior, but potentially introduced confusing error reports
if misused. We therefore settled on a simpler, though more cumbersome,
way of dealing with the issue: we introduced 'withReset0', 'withReset' -
that only worked on Signal dom a -> r -, and 'withResetProxy'. After
testing this change on larger codebases we concluded that this wasn't
satisfactory either. Upgrading to the new API still triggered confusing
errors and the new API seemed worse for no obvious reason.

This commit revisits the originally proposed solution. Once more, we
introduce type classes: HasSingleDomain and HasSpecificDomain. Both
type classes aim to tie the domain provided by the explicitely passed
clock, reset, or enable lines to the domain in the given construct.
This change differs from the original solution in that it tries to
generate clear error messages if the functions using it can't resolve
the constraint. For example, if a user tries to use 'withReset' on a
function whose first argument does not contain a domain, they will
get an error message explaining what happened:

	>>> withReset resetGen (3 :: Int)
	error:
	    • Could not find a domain in the following type:

		Int

	      This might happen if an instance for TryDomain is missing. Try to determine
	      which of the types miss an instance, and add them. Example implementations:

	       * type instance HasDomain t (MyVector n a)    = TryDomain t a
	       * type instance HasDomain t (MyCircuit dom a) = Found dom
	       * type instance HasDomain t Terminal          = NotFound

	      Alternatively, use one of the withSpecific* functions.

	      ------

	      You tried to apply an explicitely routed clock, reset, or enable line
	      to a construct with, possibly, an implicitely routed one. Clash failed to
	      unambigously determine a single domain and could therefore not route it.
	      You possibly used one of these sets of functions:

	       * with{ClockResetEnable,Clock,Reset,Enable}
	       * expose{ClockResetEnable,Clock,Reset,Enable}

	      These functions are suitable for components defined over a single domain
	      only. If you want to use multiple domains, use the following instead:

	       * withSpecific{ClockResetEnable,Clock,Reset,Enable}
	       * exposeSpecifc{ClockResetEnable,Clock,Reset,Enable}

	    • In the expression: withReset resetGen (3 :: Int)
	      In an equation for ‘it’: it = withReset resetGen (3 :: Int)

Error messages will try to be specific, although there is still a lot of
work to do.

	>>> withReset resetGen (Prelude.undefined :: Int -> Index n -> Signal XilinxSystem b)

	error:
	    • Could not determine whether the following type contained a domain:

		Index n

	      In the full type:

		Int -> Index n -> Signal XilinxSystem b

	      This might happen if an instance for TryDomain is missing. Try to determine
	      which of the types miss an instance, and add them. Example implementations:

	       * type instance HasDomain t (MyVector n a)    = TryDomain t a
	       * type instance HasDomain t (MyCircuit dom a) = Found dom
	       * type instance HasDomain t Terminal          = NotFound

	      Alternatively, use one of the withSpecific* functions.

	      ------

	      You tried to apply an explicitely routed clock, reset, or enable line
	      to a construct with, possibly, an implicitely routed one. Clash failed to
	      unambigously determine a single domain and could therefore not route it.
	      You possibly used one of these sets of functions:

	       * with{ClockResetEnable,Clock,Reset,Enable}
	       * expose{ClockResetEnable,Clock,Reset,Enable}

	      These functions are suitable for components defined over a single domain
	      only. If you want to use multiple domains, use the following instead:

	       * withSpecific{ClockResetEnable,Clock,Reset,Enable}
	       * exposeSpecifc{ClockResetEnable,Clock,Reset,Enable}

	    • When checking the inferred type [..]

Two `withReset` (and Clock/Enable) strategies now exist:

 1. `withReset`: will only work if it can determine a _single_ domain
    in a given type. Allows given reset domain to be inferred. That is
    a call such as `withReset resetGen f` will work as long as `f`
    contains a single domain.

 2. `withSpecific`: will look for a _specific_ domain in a given type.
    That is, the domain of the given clock/reset/enable must be known
    and present in the given type.

These two strategies should strike a happy balance between simple for
beginners, and fully usable for advanced users. Most importantly, it's
always safe to use.