Always inline class dictionaries
Previously, in inlineNonRep, we checked whether the type of the
scrutinee that we were inlining was of kind "Constraint".
However, somewhere along the way in one of the GHC releases,
it has become impossible to determine if something is of kind
"Constraint"; it's always of kind "TYPE" basically.

So now, we just check whether something is a class dictionary,
and if so, always inline it.