Only report NoDemand for arguments used in BlackBox
Previously, we received warnings when any of the arguments of
the Haskell implementation of a primitive had a `Absent`
demand signature; even if those arguments weren't used in the
BlackBox for the primitive either.

Now we only get a warning when an argument that's used in the
BlackBox has an `Absent` demand signature. The reason why this
situation is problematic is that expressions occurring in the
argument position of applications of this primitive are
sometimes replaced by `absentArg` by GHC; because the Haskell
implementation wasn't using those arguments anyway. However,
the BlackBox might actually want to use the literal/constant
in that "absent" position, leading to compile failures if that
argument is replaced by `absentArg`.

Fixes #320 insofar that Clash now gives warning when the
Haskell model of a primitive can lead to compile failures with
the Blackbox implementation of the model.