Refuse to compare types with stuck type families
'caseElemNonReachable' considers a branch unreachable when a branch
coerces a type to another. For example, a branch coercing `Bool ~ Int`
would be considered unreachable. It would also consider `F Char ~ Int` to
be unreachable, even if `F` was a type family eventually resolving to
`Int`. With this patch, it will try to resolve type families and refuse
to continue if given type family is stuck.

Regrettably, I haven't been able to find a small test case.