Allow partially constant structures to be constant specced
'constantSpec' used to only constant specialize on fully constant
arguments. For example, given:

    f a b = a + b + 5
    g a = f a 3

it would detect `f a 3` and rewrite `g` to:

    g a = f' 3

where `f'` would be a new top level binder with `b` substituted by `3`:

    f' a = a + 3 + 5

We could imagine a similar situation where the function takes a tuple:

    h ab = let (a, b) = ab in a + b + 5
    i a = h (a+a, 3)

'constantSpec' would conclude `(a, 3)` to be variable, and therefore do
nothing. We wouldn't realistically depend on `3 + 5` to be constant
folded, so in most cases there's no harm done. In fact, before the
release of Clash 1.0 we haven't seen it cause any issues. It turns out
that the new API can trigger these kind of issues quite easily, hence
needing a better constant folding strategy.

This patch introduces a strategy that allows partial constants to be
specialized. In short, it detects all _variable_ parts of an expression
and replaces it with a fresh binder. It then specializes and adds a
let-binding re-introducing the variable parts. So, for our previous
example, it would rewrite `i` to:

    i a = let newBndr = a + a in h' newBndr

and add a new top-level binder:

    h' aa = let (a, b) = (aa, 3) in a + b + 5

It does so recursively, therefore making sure all constants are constant
specced.