Prevent local ids shadowing global ids in 'termFreeVars'
Before this patch, local variables could accidentally shadow global
variables if they happened to have the same unique. That is, for the
lambda:

    let e = \f[l] -> f[g] f[l]

The behavior of 'termFreeVars' would be:

    termFreeVars e ~ []

Even though `f[g]` would be a free variable. This in turn would break
'globalIds', as it wouldn't consider `f[g]`.