Fix dumpVCD and add support for undefined bits
There was a bug dumpVCD where it wouldn't output all changes over time,
but only as many as the amount signals being dumped.

This was caused by an error in `toIntegers` which checks for partially
undefined values and throws an error when it finds any.

Instead of fixing that, this commit just implements support for
partially undefined values in the formatter and removes
`toIntegers`.

Fixes #788