Fix caching on CircleCI
On CircleCI each cache key can only be written to once.

Since #534 we're not running jobs on CircleCI anymore for "internal"
pull requests.
But we couldn't stop the jobs from being dispatched to CircleCI, so we
made the jobs do nothing when they are internal.

The problem with this is that the `cache_save` step still runs, and it
permanently caches the result of doing nothing.
Any future job now starts with this empty cache and has to build all
dependencies.
And when it's done it can't save its results to the cache because that
cache key already exists :(

To fix this I added a separate `cache_calc_key` step which, when we are
running tests,
calculates a key with:
  md5sum .circleci/config.yml cabal.project .ci/cabal.project.local */*.cabal
Otherwise is just a static dummy string.

I also moved the `cache_save` step to just before `run_tests`, so even
if the tests fail it will still cache the built dependencies.