Add function plurality option for blackboxes
Indicates how often a function will be instantiated in a blackbox. For
example, consider the following higher-order function that creates a tree
structure:

    fold :: (a -> a -> a) -> Vec n a -> a

In order to generate HDL for an instance of fold we need log2(n) calls
to the first argument, `a -> a -> a` (plus a few more if n is not a
power of two). Note that this only targets multiple textual instances
of the function. If you can generate the HDL using a for-loop and only
need to call ~INST once, you don't have to worry about this option. See
the blackbox for 'Clash.Sized.Vector.map' for an example of this.

Right now, option can only be generated by BlackBoxHaskell. It cannot be
used within JSON primitives. To see how to use this, see the Haskell
blackbox for 'Clash.Sized.Vector.fold'.