Fix div blackboxes
In Haskell div always rounds towards negative infinity.
But the division operator in the HDLs rounds towards zero,
in Haskell quot has this behaviour.

The old implementations of div tried to correct for by subtracting 1 from
the result when it was negative.
It led to errors when the division is exact and there was no rounding.
   8 / -2 should be -4
But the old implementation calculated it to be -5.

The new implementations use pre-compensation.
If the result will be negative it adjust the dividend by subtracting the divider
plus or minus one.
  (8 - -2 +1) / -2 = -4.5
Rounded towards zero this becomes -4.

To make sure this correction doesn't cause an overflow we have to widen the
intermediate values by one bit.