Fix (System)Verilog blackboxes for mod
Similar to div and quot, mod and rem are slightly different and
(System)Verilog only have an operator with the behaviour of rem.
So you have to correction to get from the mod result to the rem result.

The old implementation tried to do this, forgot the case of exact divisors.