X-Optimization (#918)
X-optimization has been added as a normalization pass to remove
undefined alternatives from case expressions. It is capable of
remove a case completely when only one alternative is defined, and
reducing a case expression with multiple defined branches to one
alternative per defined branch.

Netlist tests have been created to show that the previous commit
results in the case being removed when all but one cases are
undefined in an elimination. Currently it is tested that a single
defined pattern removes case completely, and multiple defined
reduce the number of alternatives to the number of defined cases.