Add 'RenderVoid' option to blackboxes
Void data (e.g., `Index 1`, `BitVector 0`, `()`) results in zero-width
vectors in HDL. HDL tools handle these constructs poorly. To work around
that 01843d3f1907111c460536ef3e37ccb679d74135 added void-filtering to
Clash. As a consequence, any functions solely generating a void
construct aren't rendered at all. Often, this is okay: functions
yielding voids are generally the result of a generalization and don't
perform any actually work when instantiated with a void. _Sometimes_
though, we would like to generate them:

 * In the case of `BiSignal`s, one would like to completely remove any
   `BiSignalOut`s from the HDL, as the corresponding `BiSignalIn` will
   already be converted to an `inout` wire. We'd still like to render
   all the blackboxes yielding a `BiSignalOut` though.

 * (SystemVerilog) Assertions don't carry any "out" signals. They
   construct a comment instructing the simulator / verification tool to
   check some property. Clash simulation can't support this though:
   functions without outputs simply do not exist. In this case, we'd
   like to mark an "assertion result" as void, so it would get filtered
   in HDL. Similar to the previous use case, we'ld like the comments
   to be rendered still. See #864 for more information.

 * Some `IO ()` actions as described in #815 must result in HDL. Like
   the previous two features, `IO ()` would get filtered though.

This commit allows blackboxes to specify whether they would like to be
rendered even if their result is void. Note that any blackboxes
specifying this must make sure that they do not assign anything to
their result - as this is zero bits, Clash won't generate any signal
declarations for it.

As a happy coincidence, this removes the need for special support for
BiSignalOut in the compiler.