[Closes #867] Actually fix doctests
It seems that because the type of element in a Vec is existentially
bound, type inference cannot tell a function that returns vector
elements is the same type. The doctests now have very ugly let
bindings in the form

  let x :: T; x = ...

to allow toplevel functions to be defined in doctests without the
test crashing. This shouldn't be an issue in local bindings within
functions (as shown by the fact functions in the vector module have
been changed to use the patterns where appropriate), so provided
functions using vectors have type annotations things should be fine.