Conditionally stop deduplication
Normally Clash converts

```
case x of
  A -> f 3 y
  B -> f x x
  C -> h x
```

to

```
let f_arg0 = case x of {A -> 3; _ -> x}
    f_arg1 = case x of {A -> y; _ -> x}
    f_out  = f f_arg0 f_arg1
in  case x of
      A -> f_out
      B -> f_out
      C -> h x
```

i.e. it deduplicates functions (and certain operators) between
case-alternatives to save on area. This comes at the cost of
multiplexers on the arguments of the deduplicated function
(or operator), which add to the propagation delay for the entire
case-expression.

So sometimes we don't want Clash to do the deduplication, i.e.
it adds to the critical path. In that case we can now specifically
tell Clash not to deduplicate:

```
case x of
  A -> noDeDup f 3 y
  B -> f x x
  C -> h x
```

where the application of `f` in the A-alternative is now
explicitly not deduplicated, and the `f` in the B-alternative
is the only remaining application of `f` it also will not be
deduplicated.

If the C-alternative also had an application of `f`, then the
applications of `f` in the B- and C-alternatives would have
been deduplicated; i.e. there would have been two applications
of `f` in total.