Fix type of (:<), add COMPLETE pragmas. Fixes 867.
The previous changes in this issue did not in practice work as
expected in all cases, as it gave some type errors for correct
definitions using (:<), and gave warnings that functions using
the pattern synonyms did not cover all cases. These have been
amended now, and the issue should be solved.

There is a new unittest for the Vector module which checks that
examples known to fail due to GADT pattern matching fail, and
that the example programs from the issue which should compile
work as desired.