Move testsuite other-modules to library (#922)
The other modules in the testsuite now form a clash-testsuite
library. This is needed so the testsuite can be included in itself
as a package when tests are run (needed to allow netlist tests to
use the runToNetlistStage function).

The cabal file has been updated while doing this, moving the minimum
version up to 2.2 (as in other subprojects) and putting common
configuration into a common stanza.