Define netlistTest for any backend (#922)
Netlist tests can now be defined over any HDl backend (as the HDL
chosen may affect the netlist due to black-boxes). An example netlist
test has been implemented which tests an identity circuit has only
a single input and output, and that the input is simply routed to
the output without anything else.