WIP - Add TickValue and CastValue
Tick and Cast (from Term) now also exist in Value. To prevent
analysis of Value's not being performed if a value is ticked, there
is collectValueTicks, which works like collectTicks.

It remains to be seen what needs to be done to make the partial
evaluator properly preserve ticks and casts.