Use insertion ordered set in 'specArgBndrsAndVars'
Before this patch it used to use a 'UniqSet'. 'UniqSet's order their contents
based on the unique of the vars they're storing. This would make the following
calls to specialise different (pseudo-code):

     specialise {f (3, x123, y456)}
     specialise {f (3, x456, y123)}

where the number after the variable label indicates a unique. This is bad,
because it breaks the caching strategy of 'specialise'.