Add "Clash.Signal.Bundle.TaggedUnit"
Helper type to emulate the "old" behavior of Bundle's unit instance. I.e.,
the instance for @Bundle ()@ used to be defined as:

    class Bundle () where
      bundle   :: () -> Signal domain ()
      unbundle :: Signal domain () -> ()

In order to have sensible type inference, the "Bundle" class specifies that
the argument type of 'bundle' should uniquely identify the result type, and
vice versa for 'unbundle'. The type signatures in the snippit above don't
though, as @()@ doesn't uniquely map to a specific domain. In other words,
'domain' should occur in both the arugment and result of both functions.

"TaggedUnit" fixes this by carrying the domain in its type.