Add "Clash.Signal.Bundle.TaggedEmptyTuple"
Helper type to emulate the "old" behavior of Bundle's unit instance. I.e.,
the instance for `Bundle ()` used to be defined as:

    class Bundle () where
      bundle   :: () -> Signal domain ()
      unbundle :: Signal domain () -> ()

In order to have sensible type inference, the "Bundle" class specifies that
the argument type of 'bundle' should uniquely identify the result type, and
vice versa for 'unbundle'. The type signatures in the snippet above don't
though, as `()` doesn't uniquely map to a specific domain. In other words,
'domain' should occur in both the argument and result of both functions.

"TaggedEmptyTuple" tackles this by carrying the domain in its type. The
'bundle' and 'unbundle' instance now looks like:

    class Bundle EmptyTuple where
      bundle   :: TaggedEmptyTuple domain -> Signal domain EmptyTuple
      unbundle :: Signal domain EmptyTuple -> TaggedEmptyTuple domain

`domain` is now mentioned both the argument and result for both 'bundle' and
'unbundle'.