Lift and inline "to-lift" let-binders one by one
In order to deal with a set of "to-lift" let-binders that form
a recursive group, e.g.

```
topEntity x y z
  = let g p q k v = k (p + q) * v
        h = ((3 :: Integer) *) . (g x y f)
        f = ((4 :: Integer) *) . (g x y h)
    in  f (h z) + h (f z)
```

The above definition remains recursive after normalization, at
the global level, but it contains no self-recursive let-bindings.
The previous implementation of the algorith would result in a
normalized set of expressions with self-recursive,
non-representable, let-bindings; which is not allowed.