Simplify ByteArrayLiteral
The ByteArrayLiteral type was a primitive vector (a bytearray with
offset and size fields included), and the only field ever used in
Clash was the ByteArray field. Consequently, the type is simply
changed to a ByteArray with no problems.

This change will also make it easier to implement ByteArray
operations in the new partial evalautor, as it means ByteArray can
be marshalled to and from Value easier.