autoDelay: autoReg without the reset
Mostly so that we don't generate reset ports for the field bits
of a `Maybe a` when the reset value is `Nothing`. Although the
field bits would subsequently be reset to undefined, there's
not guarantee that logic synthesis actually removes the reset
port on the register, it's free to leave it in and reset to
whatever it wants.

With autoDelay, we are explicit in the fact that we don't want
a reset port.