Prune Clash.Core.Util
While most functions in this module are used fairly often, there
are many which are intuitively meant to be in another module (e.g.
the functions to construct terms should be in Clash.Core.Term).

However, not all functions can be in the obvious module, as it
would result in a circular import. Rather than throw more .hs-boot
files at the problem, these are moved into a different module for
now (Clash.Core.TermInfo), so they can be handled later if the
Clash.Core namespace is tidied up.