Decouple evaluators
In the implementation of the evaluator in Clash.Core.Evaluator, there
are details about the GHC frontend which leak through. As clash-lib
is supposed to be frontend-agnostic, this is undesirable to say the
least.

The evaluator has been moved to Clash.GHC, and the evaluator for
primitives moved to Clash.GHC.Evaluator.Primitive. This frees clash-lib
to only expose a more general API. This is quite limited now, but in
the new partial evaluator will be able to contain more combinators
for default implementations of evaluators. This is not the case now,
as an evaluator expressed with small-step semantics makes this harder
(it is harder to pick apart a structure from what is being evaluated,
as each evaluation function only represents a single step in
isolation).