Define models for new partial evaluator
The new partial evaluator uses different types to distingush between
terms normalised to different forms. This makes it impossible for
an evaluator to be defined which includes syntax that no normalised
term should have (e.g. let expressions).

Conceptually, terms are evaluated to Values (WHNF) and then quoted
to obtain normal form (NF). Neutral terms are "stuck", and cannot
be reduced further. Environments provide closures, keeping track of
the types and terms that are currently in scope. Thus, the types
Value, Nf and Neutral can be converted back into terms, and Env
can be converted into a substitution.