Add CoreTest to testsuite
Similar to netlist tests added in ee5d000, core tests are now
available. These run the clash-ghc frontend as far as generating
bindings for top level definitions, giving access to the bindings
map and the TyConMap.

Core tests allow tests to be defined which inspect the result of
operating on Clash's core IR. This makes it easier to define tests
that inspect the behaviour of the partial evaluator, or
transformations in the normalizer.