Infix use of `defaultSeqX` and `deepSeqX` righ-assoc
In the register and delay primitive we were writing:

```
q `defaultSeqX` q :- rest
```

which was parsed as:

```
(q `defaultSeqX` q) :- rest
```

because of the missing fixity declaration, our intention
was the following however: !!!

```
q `defaultSeqX` (q :- rest)
```

so that the currunt value in the register was forced to
WHNF (or NF) whenever the tail of the Signal was requested
in order to prevent space-leaks. Without the fixity declaration
on `defaultSeqX`, this space-leak prevention was of course not
happening! And we were leaking space as demonstrated by
https://github.com/clash-lang/clash-compiler/issues/1256

Fixes #1256