Do not disable the monomorphism restriction
Warn when it is turned off when converting Hs to HDL. Also
enable -Wmonomorphism-restriction by default so you know when
you might want to add a type signature to prevent unexpected
monomorphism.

This commit also demonstrates the fallout of leaving the
MonomorphismRestriction enabled by default, and can be used
as a guide for fixing Clash code suffering from the same
issues once this commit lands in a released version of the
compiler.

Fixes #1270
Fixes #1345