Lazier evaluation; scoping fix; ignore lists
The partial evaluator is now lazier, deferring the evaluation of
data constructor / primitive arguments and creating thunks where
work does not need to be performed to continue.

The partial evaluator now uses an RWS monad instead of just state,
as any local changes should be performed in a Reader (meaning that
changes to local environments only persist when going down the AST
and are not present when moving back up). This fixes a hard to
identify scoping issue where some binders from letrec / case were
escaping their scope.

Ignore lists are now used to direct where inlining is not allowed
to happen. Within a particular sub-term a global binding can be
ignored, preventing it from being inlined in that branch. This
is saved in the local state when reaching a lambda / type lambda /
constructor / primitive, so that it can be restored when quoting.