toSimpleVar: Check if modifier is simple
toSimpleVar now checks if a modifier is simple in an identifier
(only identifiers with no modifier are simple; this can be changed
later to something more liberal). If it is not simple, a separate
signal is created in the output HDL to hold the modified expression.

(cherry picked from commit e1d880102ce1b2cd2b262ce0d20927382f539976)