Store whether bindings are primitive unfoldings
In the partial evaluator, it is important to be able to fallback
to the core unfolding for a primitive without an explicit rule.
However, these primitives are marked as NOINLINE in source, meaning
the evaluator will refuse to inline them.

Rather than ignoring INLINE pragmas, we track whether a binding
refers to a primitive with a known implementation. If this is true,
the evaluator can use the core unfolding. The change to the
evaluator is not made in this commit.