Add a high-speed version of AutoReg
This still splits registers of products into products of registers,
but it no longer uses the tag bit to control the enable lines
of the registers holding the value bits.

This decreases the logic depth, thus shortening the potential
critical path, at the cost of power, since the fields bit can
freely toggle.