Move work free checks to Clash.Rewrite.WorkFree
Tests for work free / work freeish now belong in a separate module.
This means they can be used in the partial evaluator branch without
introducing a circular dependency in imports.

In the process of changing isWorkFreeIsh to not use RewriteMonad
it became a pure function (as it does not cache like isWorkFree).