Non TH 'fromList' and 'unsafeFromList' for Vec.
These functions allow Vectors to be created from a list without needing to
use template haskell, which is not always desirable. The unsafe version of the
function does not compare the length of the list to the desired length of the
vector, either truncating or padding with undefined if the lengths differ.