Remove absentError from primops which are values
absentError was added to the lists of primitives which are
undefined and primitives which are values. These two lists should be
disjoint, as when unwinding primitives which are values are preserved
and primtives which are undefined are replaced with
Clash.Transformations.undefined.

Primitives which are values are checked before primitives which are
undefined, meaning currently absentError will be preserved in the
output from compile time eval, despite being undefined. As other GHC
error primitives are not treated as values, the correct behaviour
should be to treat it is undefined.