Change work free checks
When checking applications where the LHS is a variable, the function
checked if it was polymorphic / function type before checking if it
is a local id. This seems reasonable at first glance, but in the
partial evaluator this means that we can end up with core like

  \arg ->
    let f0 = \x -> x in
    let f1 = f0 in
    ...
    let fn = fm in
    fn arg

as the term bound to f0 is a function type, it is preserved. However,
the let bindings which are simply local variables could be inlined
further producing

  \arg ->
    let f0 = \x -> x in
    f0 arg

which significantly reduces the work needed after PE in the rewriting
pipeline. In the example program at

  https://github.com/clash-lang/clash-compiler/issues/1536#issuecomment-775918287

this reduces the number of transformations from 17 to 5.