hzToPeriod/periodToHz: Use Ratio Natural
This change does not significantly affect the generated circuit,
although the change is observable in simulation.

We would like `hzToPeriod` to satisfy the following:

- Results should be intuitive
- Results should match a type-level equivalent as proposed in issue
  #1253
  https://github.com/clash-lang/clash-compiler/issues/1253

The former can be done with `Double`, by rounding to the nearest
integer. This means that if there is an exact integer solution, we give
that. Before this commit we had:

>>> hzToPeriod 100_000
10000001

instead of the intuitive result that a 100 kHz clock corresponds to a
10 µs period. The off-by-one has no significant consequences for the
circuit, clocks are inherently not precise. It is just not intuitive.

But if we also want the result to match a simply expressed type-level
equivalent, this is not enough, because the slight errors introduced by
`Double` might cause a fractional part of 0.5 to be rounded either up or
down seemingly randomly.

So the most straightforward fix is to use `Ratio Natural` and `floor`.
If the result is an integral number of picoseconds, it will give that,
and for roundings, it will match the behaviour of the straightforward
type-level equivalent of:

type HzToPeriod (freq :: Nat) = 1_000_000_000_000 `Div` freq