Remove EXPERIMENTAL_EVALUATOR
The partial evaluator should not replace the old evaluator
everywhere it is currently used, rather it should _only_ be used
for partial evaluation. This is because the partial evaluator may
not result in something that still repsects invariants (e.g. it
may not preserve invariants expected to hold after ANF).

Still TODO: It should probably be possible to selectively turn PE
on or off by a flag (i.e. -fclash-partial-evaluator). I'm not sure
how to do this in a way where the testsuite is still useful...