Make InScopeSet part of LocalEnv
It is common in PE for the same global identifier to be inlined
multiple times, and thus the same uniques to be deshadowed. Keeping
the InScopeSet global means that the number of iterations for
uniqAway can grow quickly, and running with +debug gives the
error "uniqAway: loop".

As uniques only need to be unique for variables in scope, the
InScopeSet can be made part of the local environment, meaning
variables in different case alternatives etc. can have the same
unique (as they will never shadow each other).