Add isExpandable / expandableValue
When checking if terms can be inlined into a case expression, it
can be more useful to check if they are expandable as opposed to
work free. An expandable term is one where inlining may lead to
a new term which case-of-constructor can fire on.

Work free and expandable checks are now simpler, using asTerm and
deferring to the implementation on Term. This helps to avoid
unexpected differences between checking Term and Value.