Use effective number of CPUs on CI, not available cores
Before this patch, CI would use `nproc` to determine the number of CPUs
in a Docker instance. This only works when the docker container is
pinned to certain cpus, not if it's just rate limited. This script
detects that rate limiting, allowing our GitLab runners to unpin the
runners. This would allow us to "oversell" GitLab runners, hopefully
reducing the idle time imposed by single-threaded workloads.