Correct `IsActiveEnable` in `renderElem` (#1788)
Fixes #1786.

Some functionality is added: when `~ISACTIVEENABLE` is used on a signal
of type `Signal dom Bool` and it is always true, `~ISACTIVEENABLE` now
evaluates to true as well. Before this commit, it was not an error to
use `~ISACTIVEENABLE` on a `Signal dom Bool`, but it would not be
checked whether it was always true or not.

(cherry picked from commit 70dd6de67420e0acc34e1d0a28ca23fbc2ee79f7)