Add Clash.Normalize.Primitives
A module has been created which exports the special primitives
created during normalization. The eqInt primitive has not been
defined here, as it is defined identically to GHC.Classes.eqInt.
While the clash-specific primitive is more general, it has yet
to be needed as the only supported frontend is GHC. This primitive
will be removed later.