Fix floating point literal alpha equivalence
When we switched to using `Float` and `Double` in our
`Clash.Core.Literal` data type in #1811, our `Eq` and `Hashable`
instances for `Literal` were made incorrect. We need floating point
literals `x` and `y` to be equal iff `pack x == pack y`.

We already assume that `Float` is single-precision IEEE-754 and
`Double` is double-precision IEEE-754, so we might as well represent
them as 32/64-bits words to fix equality and hashing.