Change negate implementation for Index
The implementation of negate for Index was implemented as

  maxBound -# i

which is not really a negation, and produced some surprising
results, e.g. negate 0 == maxBound. A more sensible implementation
is treating negate as equivalent to

  satSub SatWrap 0 i

which still produces somewhat surprising values (on account of
every argument other than 0 underflowing) but is more in line with
what is implemented by other numeric types.