Fix bitrot in profiling code
The clash-profile-normalization-run broke at run time because
the deserialized data had a different type signature then the
serialiazed data in clash-profile-normalization-prepare.
This was broken by the refactor in:
https://github.com/clash-lang/clash-compiler/pull/1057/commits/df551ecdd65c5c19474fc71108a0ed41d95bd0ab

There is now one type alias shared by both, resulting in compile errors
in both files in future refactors like this.

Also I've remove the unused tuple element leading to this mess.

And while it wasn't broken, apply the same security measures to clash-profile-netlist-*