Upgrade CI docker image
The CI docker image is now multi-stage, and uses buildkit on compatible
machines to build stages in parallel. This means that changes to the
dockerfile (like updating a dependency build from source) will now
no longer force any other dependencies to be rebuilt.

As a bonus, the final stage now only has dependencies needed to build
and run clash-testsuite, meaning the image is ~0.7GB smaller.