Remove temporaryDirectory hack in testsuite
Output and Library tests now use the same temporary directory
creation as the normal runTest type tests, instead of using the
old temporaryDirectory hack.

As a side-effect of this, the different parts of a test are now
different types which implement IsTest. ClashTest is now
ClashGenTest, as it was used to generate HDL with Clash, and there
is also ClashBinaryTest for making and executing a normal Haskell
executable.