Better documentation for ANF
Issue #489 calls for better documenting ANF and ensuring that the
result of the transformation is implemented as described. The first
part of that is done here, documenting types which describe the
intended result of ANF and list the invariants which Clash expects
to hold in the result of ANF.

The second part, ensuring the result matches this spec, is not
performed. Changing the existing ANF implementation to use these
types seems to be non-trivial, and as running clash in debug mode
suggests that the transformation respects these invariants the
effort to reward ratio seems too low for now.