VHDL: don't overflow natural range in shift/rotate
Haskell `shift{R;L}` and `rotate{R;L}` take an `Int` as the
shift amount. In VHDL, `shift_{left;right}` and
`rotate_{left;right}` take a `natural` as a shift amount.
On 64 bit systems Haskell's `maxBound :: Int` is larger
than VHDL's `natural'high` (in (nearly) all VHDL simulators).

So for `shift_{left;right}` we saturate the shift amount to
`natural'high`, and for `rotate_{left;right}` we do the rotate amount
modulo the word width, all in order for the VHDL simulator not to error
out with an out-of-bounds exception when the shift/rotate amount is
larger than `natural'high`.