VHDL: don't overflow integer range in RAM/ROM
Most RAM and ROM functions take an `Int` as either a read or
write index. On 64-bit systems `maxBound :: Int` and `minBound :: Int`
exceed VHDL's `integer'high` and `integer'low` (on (nearly) all
VHDL simulators). This poses an issue, because the blackbox
implementations use `to_integer` to convert the `signed (63 downto 0)`
(then translation for `Int`) to VHDL's `integer`. So any value
falling outside of the range would trigger an overflow execption
in the VHDL simulator.

So we now masks all indexes to fit the `integer'low to integer'high`
range.