Fix shiftL heap overflow
For `BitVector`, `Signed`, `Unsigned`, `Index` and `Fixed`, if shiftL
was invoked with a really large shift amount, this could lead to a heap
overflow.

The shift was performed using `Integer` first, and only then brought
back to the proper word length. This built an extremely large `Integer`
for a moment.

Since we opted to have a shift by more than the word length result in
the number 0, that case is explicitly handled to avoid building a large
`Integer`.

Additionally, I noticed some `shift` error messages included the actual
shift amount and others did not. All `shift` and `rotate` error messages
now include the actual problematic amount in the error message.